+++ src/gray8.rs
@@ -7,7 +7,7 @@ use crate::pixel::{PixFmt, lerp_u8};
 /// 8-bit grayscale [pixel format](trait.PixFmt.html).
 ///
 /// This pixel format is for 8-bit grayscale with no alpha channel.
-#[derive(Clone, Copy, Debug, Default)]
+#[derive(Clone, Copy, Debug, Default, PartialEq)]
 #[repr(C)]
 pub struct Gray8 {
     value: u8,