+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 * Raster implements Clone and Debug
 * Raster::set_rect
+* Implemented PartialEq for alpha8, gray8, rgb8 and rgba8
 ### Changed
 * Updated to rust 2018 edition